rename om package as gom
