+++ intergo_test.go
@@ -8,7 +8,6 @@ import (
 
 	"github.com/mathetake/intergo"
 	"github.com/mathetake/intergo/bm"
-	"github.com/mathetake/intergo/om"
 	"github.com/mathetake/intergo/tdm"
 )
 
@@ -55,7 +54,7 @@ func BenchmarkMultileaving(b *testing.B) {
 
 			for _, samplingSize := range []int{2, 10, 50, 100} {
 				b.Run(fmt.Sprintf("[[%d-th bench on Optimized Multileaving with sampling size: %d]]", n, samplingSize), func(b *testing.B) {
-					benchmarkInputNum(fxx, inputRankingNum, &om.OptimizedMultiLeaving{
+					benchmarkInputNum(fxx, inputRankingNum, &gom.OptimizedMultiLeaving{
 						NumSampling: samplingSize, CreditLabel: 0, Alpha: 0,
 					}, b)
 				})
